fix: hardness truncation and attribute listener vs packet adapter conflict
1. Fix double-to-long truncation in getPeriod() for NoteBlock, StringBlock,
   and ChorusBlock listeners. getHardness() returns double but was assigned
   to a long variable, silently truncating fractional values.

2. Only register CustomBlockMiningListener when the packet adapter is NOT
   enabled. When ProtocolLib/PacketEvents is active, BreakerSystem cancels
   START_DIGGING packets before BlockDamageEvent fires, so the attribute-
   based listener never receives events. The BreakerSystem timer-based
   approach handles custom hardness in that case.

+++ core/src/main/java/io/th0rgal/oraxen/mechanics/provided/gameplay/chorusblock/ChorusBlockMechanicListener.java
@@ -68,7 +68,7 @@ public class ChorusBlockMechanicListener implements Listener {
             public long getPeriod(final Player player, final Block block, final ItemStack tool) {
                 final ChorusBlockMechanic mechanic = OraxenBlocks.getChorusMechanic(block);
                 if (mechanic == null) return 0;
-                final long hardness = mechanic.getHardness();
+                final double hardness = mechanic.getHardness();
                 double modifier = 1;
                 if (mechanic.getDrop().canDrop(tool)) {
                     modifier *= 0.4;
